requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-crispy-forms==1.14.0
django-environ==0.10.0
django-postgres-extra==2.0.8
django-sass-processor==1.2.1
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
